third_party/blink/web_tests/platform/generic/fast/forms/textarea/textarea-input-event-expected.txt

Test for https://bugs.webkit.org/show_bug.cgi?id=15189


On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS ta.value is "x"
PASS ta.value is "x"
PASS handleTextareaOnInputCallCount is 1
PASS handleDivOnInputCallCount is 1
PASS handleTextareaOnInputCallCount is 1
PASS handleDivOnInputCallCount is 1
PASS successfullyParsed is true

TEST COMPLETE
